Lifted state up.
Order now uses the new DBData class to load database info just once and pass it down to the lower components.

ItemList now keeps track of item info and passes it down to Item objects.  (This is will need to be lifted again up into order.)

Next up: make sure that ItemData object stays in sync with the Selectors.  Right now, for example, if I change the item type, the stored item_id does not update.